pyproject.toml
[build-system]
requires = [
    "setuptools",
    "setuptools-git-versioning"
]
build-backend = "setuptools.build_meta"

[tool.setuptools-git-versioning]
enabled = true
dev_template = "{tag}"

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "asemd"
dynamic = ["version"]
maintainers = [
    {name = "Viktor Svahn", email="viktor.svahn@kemi.uu.se"}
]
requires-python = ">=3.8"
description = "CLI for the atomic simulation environment (ASE)."
readme = {file = "README.md", content-type = "text/markdown"}
license = {file = "LICENSE"}
dependencies = [
    "numpy",
    "ase",
    "pandas",
    "pyyaml"
]

[project.urls]
repository = "https://github.com/viktorsvahn/asemd"

[project.scripts]
asemd = "asemd.__main__:main"
aseconf = "aseconf.__main__:main"